Feito calculo do valor corrigido

+++ src/main/java/com/deliverit/cpagar/model/ContasPagarModel.java
@@ -24,7 +24,7 @@ public class ContasPagarModel {
     private int id;
 
     @NonNull
-    private BigDecimal valorOriginal;
+    private double valorOriginal;
 
     @NonNull
     @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
